em da fix

+++ ex_4/product_management/src/main/java/com/example/product_management/controller/ProductController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 
 @Controller
@@ -26,21 +27,32 @@ public class ProductController {
     }
 
     @PostMapping("/create")
-    public String create(Product product) {
+    public String create(@ModelAttribute Product product,RedirectAttributes redirectAttributes) {
         productService.create(product);
+        redirectAttributes.addFlashAttribute("msg", "Create Successful!");
         return "redirect:/";
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable int id) {
-        productService.delete(id);
+    public String delete(@PathVariable int id, RedirectAttributes redirectAttributes) {
+        if(productService.findByID(id)==null){
+            redirectAttributes.addFlashAttribute("msg","Delete Failed!");
+        }else {
+            productService.delete(id);
+            redirectAttributes.addFlashAttribute("msg", "Delete Successful!");
+        }
         return "redirect:/";
     }
 
     @GetMapping("/detail/{id}")
-    public String detail(@PathVariable int id,Model model){
-        model.addAttribute("product_detail",productService.findByID(id));
-        return "/detail";
+    public String detail(@PathVariable int id,Model model,RedirectAttributes redirectAttributes){
+        if(productService.findByID(id)==null){
+            redirectAttributes.addFlashAttribute("msg","This product not found!");
+            return "redirect:/";
+        }else {
+            model.addAttribute("product_detail", productService.findByID(id));
+            return "/detail";
+        }
     }
 
     @GetMapping("/showUpdateForm/{id}")
@@ -50,14 +62,23 @@ public class ProductController {
     }
 
     @PostMapping("/update")
-    public String update(Product product){
-        productService.update(product);
+    public String update(@ModelAttribute Product product,RedirectAttributes redirectAttributes){
+        if(productService.findByID(product.getId())==null){
+            redirectAttributes.addFlashAttribute("msg","This product not found!");
+        }else {
+            productService.update(product);
+            redirectAttributes.addFlashAttribute("msg","Update Successful!");
+        }
         return "redirect:/";
     }
 
     @GetMapping("/search")
-    public String search(@RequestParam String search_name, Model model){
-        model.addAttribute("products",productService.search(search_name));
+    public String search(@RequestParam String search_name, Model model,RedirectAttributes redirectAttributes){
+        if(productService.search(search_name).size()==0){
+            redirectAttributes.addFlashAttribute("msg","This product not found!");
+        }else {
+            model.addAttribute("products", productService.search(search_name));
+        }
         return "/display";
     }
